### 2025-05-17 #8221 @editorialbot
[PRE REVIEW]: Syd: A package for making interactive data visualizations in python
**Submitting author:** @landoskape (<a href="http://orcid.org/0000-0001-9105-1636">Andrew Landau</a>)
**Repository:** https://github.com/landoskape/syd
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.3
