### 2025-05-19 #8221 @landoskape
Possible Reviewers:

domoritz (author of altair, very similar)
dnszafir, terrytangyuan, kellieotto (reviewers of altair)
jkrumbiegel (author of makie)
fverdugo, gaelforget (reviewers of makie)
