### 2025-06-28 landoskape/syd#12 @ZeitgeberH
[JOSS REVIEW]  comments from Z.
Nice work! The notebook view works like a charm!

About the package:
- [x] The web-browser view has some limitations.  It seems it only supports one visualization (served at port 5000), either from notebook or script. If one tries to spin-up a different one, it won't replace the first one unless the latter's kernel was closed before hand. This isn't what author claimed currently, but an explicit documentation of it may avoid some confusion.  Alternatively, use different ports for different visualizations.
- [x] In example 2 (both 2a and 2b),the callbacks for "update_grid" and "update_legend" are somehow tied up:  Updating one will automatically update the other.

- [x] Add a "\__init__.py" in the **tests** folder for pytest to run the whole folder in one go.

About the paper

- [x] A brief comparison with other packages (like Panel, Plotly for example ) which have similar functionalities should satisfy the requirement for the "State of the field"

Minor comment:
In the README's QuickStart section, the description about env (env="notebook" or :browser) seems outdated.

[link back to review](https://github.com/openjournals/joss-reviews/issues/8447)
