### 2025-06-28 landoskape/syd#12 @ZeitgeberH
The following comments are just suggestions as a potential user not for reviews.
-  Support seaborn would be great feature; Not just aesthetics, but also for direct zoom-in/out/pan features which are probably the most useful visualization features in a notebook.
- ROI annotations: for imaging folks, this would be a cool feature to draw ROIs or dynamically loading traces within the ROIs.
- Examples of viewing video or image stack?
