### 2025-07-06 8931d21
add control over updating plot
