### 2025-07-08 a9b442d
more examples and explanations
