### 2025-07-08 20294c4
Update paper.md
