### 2025-07-08 2df330e
Update paper.md
