### 2025-07-08 landoskape/syd#12 @landoskape
>   [ ]  The web-browser view has some limitations.  It seems it only supports one visualization (served at port 5000), either from notebook or script. If one tries to spin-up a different one, it won't replace the first one unless the latter's kernel was closed before hand. This isn't what author claimed currently, but an explicit documentation of it may avoid some confusion.  Alternatively, use different ports for different visualizations.

Thank you for catching this! There was a bug in the code for finding the next available port that I fixed. It can spin-up as many as you want now.

>   [ ]  In example 2 (both 2a and 2b),the callbacks for "update_grid" and "update_legend" are somehow tied up:  Updating one will automatically update the other.

Ah, this was actually what I intended to show how you can use custom callbacks, but I agree it's confusing. I separated their functions and demonstrate user-defined callbacks in a different way.

>   [ ]  Add a "__init__.py" in the **tests** folder for pytest to run the whole folder in one go.

Done.

> * [ ]  A brief comparison with other packages (like Panel, Plotly for example ) which have similar functionalities should satisfy the requirement for the "State of the field"

I added a section called "Why choose Syd" which should address this. I view Syd as a complement to existing fantastic packages, not a replacement. Syd's usefulness is in it's simplicity.. many other packages can do what Syd can do, but either with more boilerplate, higher learning curves, they don't work in notebooks, or are simply so powerful that the documentation is often focused on other things which makes it harder to learn this specific function. LMK if you agree with the section in the paper, I believe you should be able to see the pdf artifact from the "paper" action?

> Minor comment: In the README's QuickStart section, the description about env (env="notebook" or :browser) seems outdated.

Oops, fixed!

> [link back to review](https://github.com/openjournals/joss-reviews/issues/8447)
