### 2025-07-12 landoskape/syd#12 @ZeitgeberH
You do have a "port" parameter in the viewer.share() function. It was not mentioned in the documentation. Maybe add a brief comment to the quick start guide, which may be helpful is someone want to use a specific port for some reason.
