### 2025-07-12 #8447 @ZeitgeberH
My review of **syd** package has completed.  It is a simple & versatile toolkit for busy scientists who want to quickly get an interactive visualization, both locally or sharing via network. I recommend the package be accepted for publication.
