### 2025-08-26 #8447 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
